move flow types to schema types
