+++ tests/test_schema_manager.py
@@ -4,7 +4,7 @@ from contextlib import contextmanager
 from typing import Generator
 from pathlib import Path
 from barfi.flow.schema.manage import SchemaManager
-from barfi.flow.flow.types import FlowSchema, FlowViewport, SCHEMA_VERSION
+from barfi.flow.schema.types import FlowSchema, FlowViewport, SCHEMA_VERSION
 
 
 @contextmanager
